+++ backend/core/taps/ripe_ris.py
@@ -4,7 +4,7 @@ from ipaddress import ip_network as str2ip
 import json
 from kombu import Connection, Producer, Exchange
 import os
-from utils import mformat_validator, normalize_msg_path, key_generator, RABBITMQ_HOST, get_logger, is_subnet_of
+from utils import mformat_validator, normalize_msg_path, key_generator, RABBITMQ_URI, get_logger, is_subnet_of
 import websocket
 
 log = get_logger()
@@ -162,7 +162,7 @@ if __name__ == '__main__':
     host = args.host
 
     try:
-        with Connection(RABBITMQ_HOST) as connection:
+        with Connection(RABBITMQ_URI) as connection:
             parse_ripe_ris(connection, prefix, host)
     except Exception:
         log.exception('exception')